Refactor
Refactored the code to get rid of old stuff that was not being used or called. Improved Job code with new constructor, comments, kind of new way to do haul code. Resources now check to see if there are open stockpiles - need to optimize the way I am doing it